requirements.txt
streamlit
pandas
numpy
matplotlib
Pillow
requests